pyproject.toml
[tool.poetry]
name = "Telegram bot"
version = "0.1.0"
description = ""
authors = ["Sergey Pavlov <self@sergeypavlov.dev>"]

[tool.poetry.dependencies]
python = "3.9.7"
requests = "2.26.0"
python-telegram-bot = "13.7"

[tool.poetry.dev-dependencies]
mypy = "0.910"
flake8 = "3.9.2"
types-requests = "2.25.9"

[tool.poetry-exec-plugin.commands]